fix: remove storage class specification
Remove the hard standard-rwo storage class from PVC templates. This is
GKE-specific, and will break installs on clusters that lack an
equivalent.

Specifying no storage class will use the cluster's default storage
class. This is standard-rwo on Autopilot-managed GKE.

If we wish to re-add this, it should be exposed in values.yaml and
default to no value, as there is not a guaranteed available class other
than the local default.